+++ src/modules/commands/googleit.ts
@@ -16,7 +16,7 @@ export default new Command()
     .use(context => {
         const { message } = context;
 
-        message.delete();
+        if (message.guild !== null) message.delete();
 
         const attachment = new MessageAttachment(
             `src/common/images/googleit.png`
