+++ src/modules/commands/setpinchannel.ts
@@ -42,7 +42,7 @@ export default new Command()
         const { message } = context;
         const { args } = context.state;
 
-        message.delete();
+        if (message.guild !== null) message.delete();
 
         if (!message.member!.hasPermission("ADMINISTRATOR")) {
             return message.channel.send(
